+++ NEWS.md
@@ -1,5 +1,9 @@
 # testthat 0.X
 
+## New features
+* `skip_on_travis()` allows you to skip tests when run on Travis CI.
+  (Thanks to @mllg)
+
 ## Removed functions
 
 * `colourise()` was removed. (Colour is still supported, via the `crayon` package.)